Extracted path planner logic to separate class
