+++ Zoop/SmallGridZoopPlan.cs
@@ -1,0 +1,10 @@
+using System.Collections.Generic;
+
+namespace ZoopMod.Zoop;
+
+internal sealed class SmallGridZoopPlan(List<ZoopSegment> segments, bool isSinglePlacement)
+{
+
+  public List<ZoopSegment> Segments { get; } = segments;
+  public bool IsSinglePlacement { get; } = isSinglePlacement;
+}
